ScaffoldingCode/requirements.txt
kafka-python
numpy
Pillow
tensorflow
opencv-python
                  
